添加新的配置文件和数据统计，更新特征提取器参数，修改推理脚本以支持新的输入格式

+++ optispeech/dataset/feature_extractors/__init__.py
@@ -195,6 +195,6 @@ class CommonFeatureExtractor(FeatureExtractor):
         )
 
         magnitudes = torch.sqrt(spec.pow(2).sum(-1) + (1e-9))
-        spec = torch.matmul(self.mel_basis[mel_basis_key], magnitudes)
+        spec = torch.matmul(self.mel_basis[mel_basis_key].float(), magnitudes.float())
         spec = spectral_normalize_torch(spec)
         return spec.squeeze().cpu().numpy()
